Compilation fixes

+++ cores/esp32/WMath.cpp
@@ -41,11 +41,11 @@ long random(long howbig)
     uint32_t x = esp_random();
     uint64_t m = uint64_t(x) * uint64_t(howbig);
     uint32_t l = uint32_t(m);
-    if (l < howbig) {
+    if (l < (uint64_t)howbig) {
         uint32_t t = -howbig;
-        if (t >= howbig) {
+        if (t >= (uint64_t)howbig) {
             t -= howbig;
-            if (t >= howbig)
+            if (t >= (uint64_t)howbig)
                 t %= howbig;
         }
         while (l < t) {
